// src/app/components/model/item.model.ts
export interface Item {
    id: string
    name: string
    quantity: number;
    price: number;
    type: string;
    description: string;
}

// private UUID id;

// @NotBlank
// private String name;

// private Long quantity;

// private Long price;

// private String type;

// private String description;